refactor(refs T30969): move DpSearchField to components root
