+++ src/components/index.js
@@ -9,6 +9,7 @@ import DpInput from './DpInput/DpInput'
 import DpLabel from './DpLabel/DpLabel'
 import DpLoading from './DpLoading/DpLoading'
 import DpRadio from './DpRadio/DpRadio'
+import DpSearchField from './DpSearchField/DpSearchField'
 import DpSelect from './DpSelect/DpSelect'
 import DpTextArea from './DpTextArea/DpTextArea'
 import {
@@ -42,7 +43,6 @@ import {
   DpProgressBar,
   DpRegisterFlyout,
   DpResettableInput,
-  DpSearchField,
   DpSelectPageItemCount,
   DpSkeletonBox,
   DpSlidebar,
